feat(cpp): Can use existing code to build P2WPKH

+++ src/Bitcoin/Signer.cpp
@@ -26,7 +26,7 @@ Proto::SigningOutput Signer::sign(const Proto::SigningInput& input, std::optiona
     if (input.is_it_brc_operation()) {
         auto serializedInput = data(input.SerializeAsString());
         Rust::CByteArrayWrapper res = Rust::tw_taproot_build_and_sign_transaction(serializedInput.data(), serializedInput.size());
-        output.ParseFromArray(res.data.data(), res.data.size());
+        output.ParseFromArray(res.data.data(), static_cast<int>(res.data.size()));
         return output;
     }
     auto result = TransactionSigner<Transaction, TransactionBuilder>::sign(input, false, optionalExternalSigs);
